refactor of errors for blockorder
