+++ broker-daemon/broker-rpc/order-service/cancel-block-order.spec.js
@@ -1,6 +1,7 @@
 const path = require('path')
 const { expect, rewire, sinon } = require('test/test-helper')
-const { BlockOrder } = require('../../models')
+
+const { BlockOrderNotFoundError } = require('../../models/errors')
 
 const cancelBlockOrder = rewire(path.resolve(__dirname, 'cancel-block-order'))
 
@@ -8,8 +9,6 @@ describe('cancelBlockOrder', () => {
   let PublicError
   let blockOrderWorker
   let blockOrder
-
-  const { BlockOrderNotFoundError } = BlockOrder.ERRORS
 
   beforeEach(() => {
     PublicError = cancelBlockOrder.__get__('PublicError')
