+++ broker-daemon/broker-rpc/order-service/get-block-order.spec.js
@@ -1,6 +1,7 @@
 const path = require('path')
 const { expect, rewire, sinon } = require('test/test-helper')
-const { BlockOrder } = require('../../models')
+
+const { BlockOrderNotFoundError } = require('../../models/errors')
 
 const getBlockOrder = rewire(path.resolve(__dirname, 'get-block-order'))
 
@@ -9,8 +10,6 @@ describe('getBlockOrder', () => {
   let GetBlockOrderResponse
   let blockOrderWorker
   let blockOrder
-
-  const { BlockOrderNotFoundError } = BlockOrder.ERRORS
 
   beforeEach(() => {
     PublicError = getBlockOrder.__get__('PublicError')
