+++ broker-daemon/models/block-order.js
@@ -2,21 +2,7 @@ const { promisify } = require('util')
 
 const { Big } = require('../utils')
 const CONFIG = require('../config')
-
-class BlockOrderNotFoundError extends Error {
-  /**
-   * @param {String} id - a blockorder id
-   * @param {Error} err
-   */
-  constructor (id, err) {
-    super(`Block Order with ID ${id} was not found.`, err)
-
-    this.name = this.constructor.name
-    this.notFound = true
-
-    if (err) this.stack = err.stack
-  }
-}
+const { BlockOrderNotFoundError } = require('./errors')
 
 /**
  * @class Model representing Block Orders
