+++ broker-daemon/models/errors/index.js
@@ -1,0 +1,5 @@
+const BlockOrderNotFoundError = require('./block-order-not-found-error')
+
+module.exports = {
+  BlockOrderNotFoundError
+}
